Cleaned up use of debug/testDebug Streams
debug is now the standard Stream for debugging messages which may be
sent over Zigbee.  localDebug always sends messages locally and does not
involve Zigbee because those messages are often for debugging Zigbee
itself.  serial now is used by parseSerial when just talking to the
local serial port for control.  Currently it's the same as localDebug.